requirements.txt
django
djangorestframework
python-dotenv
PyMySQL
mysqlclient
auth
djangorestframework-simplejwt
django-filter
django-cors-headers
Pillow
